# stackapp_openapi.yaml
openapi: 3.1.0
info:
  title: StackApp API
  description: |
    # 🎤 StackApp - The Black Community's Financial Powerhouse
    
    **"Stack Your Bread, Stack Your Future"**
    
    StackApp is a revolutionary financial platform designed specifically for the Black community, combining authentic cultural voice with cutting-edge AI technology to provide personalized financial guidance and wealth-building strategies.
    
    ## 🆓 FREE TIER - Qwen AI Model
    - **No API keys required**
    - **Real AI responses** using Qwen/Qwen2.5-7B-Instruct
    - **Authentic Stack Master personality** - culturally aware financial coaching
    - **Core financial advice** - investment, credit, savings guidance
    - **Community features** - challenges, success stories, education
    
    ## 💎 PREMIUM TIER - FinRobot Agents ($9.99 Beta)
    - **Advanced FinRobot agents** - professional-grade financial analysis
    - **Real-time market data** - live financial information
    - **Personalized strategies** - custom investment plans
    - **Priority support** - dedicated assistance
    
    ## 🎯 Revenue Model
    - **Free users**: Qwen AI model with full Stack Master personality
    - **Premium users**: FinRobot agents with advanced analysis capabilities
    - **Beta pricing**: $9.99/month for early adopters
    
    ## 🚀 Getting Started
    1. **No registration required** for free tier
    2. **Start chatting** with The Stack Master immediately
...
    
    ## 🎤 The Stack Master
...
    
...
  contact:
...
servers:
...
paths:
...
  /stack-master/chat:
    post:
      tags:
        - Stack Master
      summary: Chat with The Stack Master
      description: |
        Chat with The Stack Master AI Coach using the free Qwen model. 
        Get authentic, culturally-aware financial advice with no API keys required.
        
        **Free Tier Features:**
        - Real AI responses using Qwen/Qwen2.5-7B-Instruct
        - Authentic Black community voice and personality
        - Personalized financial advice
        - Actionable steps and motivational messages
      operationId: chatWithStackMaster
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StackMasterMessage'
            examples:
              investment_question:
                summary: Investment advice request
                value:
                  user_id: "user123"
                  message: "I want to start investing but I only have $50 a month. What should I do?"
                  context:
                    income: 3000
                    savings: 200
                    risk_tolerance: "conservative"
              credit_question:
                summary: Credit building request
                value:
                  user_id: "user456"
                  message: "My credit score is 580. How can I improve it quickly?"
                  context:
                    credit_score: 580
                    monthly_income: 2500
                    current_debt: 5000
              general_question:
                summary: General financial advice
                value:
                  user_id: "user789"
                  message: "I'm 25 and want to build wealth. Where do I start?"
                  context:
                    age: 25
                    income: 4000
                    savings: 1000
      responses:
        '200':
          description: Successful response from The Stack Master
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StackMasterResponse'
              examples:
                investment_response:
                  summary: Investment advice response
                  value:
                    response: "Real talk - let's break this down. First step is tracking every dollar. You can't stack what you can't see. Start with the 50/30/20 rule: 50% needs, 30% wants, 20% wealth building. Investing $50 a month can turn into $50k over time. Compound interest is how the rich stay rich. Time to join the club."
                    advice_type: "investment"
                    actionable_steps:
                      - "Start with $25/month in an index fund"
                      - "Research low-cost ETFs like VTI or SPY"
                      - "Set up automatic transfers to your investment account"
                    motivational_message: "Your future self will thank you for starting today!"
                credit_response:
                  summary: Credit building response
                  value:
                    response: "Your credit score is your power score. Every payment on time is money in the bank. Let's get you to 750+ and unlock real wealth opportunities. Credit is power - let's build yours!"
                    advice_type: "credit"
                    actionable_steps:
                      - "Pay all bills on time, every time"
                      - "Keep credit card balances below 30% of limit"
                      - "Check your credit report monthly"
                    motivational_message: "Every payment on time is money in your pocket!"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                server_error:
                  summary: Server error
                  value:
                    detail: "Error chatting with Stack Master: Connection timeout"
  /stack-master/analyze-stack:
    post:
      tags:
        - Stack Master
      summary: Analyze financial stack
      description: |
        Get a comprehensive analysis of your current financial situation with personalized recommendations.
        Uses the free Qwen model to provide culturally-aware financial guidance.
      operationId: analyzeFinancialStack
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StackAnalysisRequest'
            examples:
              basic_analysis:
                summary: Basic financial analysis
                value:
                  user_id: "user123"
                  current_income: 3000
                  current_savings: 500
                  monthly_expenses: 2200
                  credit_score: 650
                  investment_goals: ["emergency fund", "retirement"]
                  risk_tolerance: "moderate"
              minimal_analysis:
                summary: Minimal information analysis
                value:
                  user_id: "user456"
                  current_income: 2500
                  risk_tolerance: "conservative"
      responses:
        '200':
          description: Financial analysis completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                    example: "user123"
                  analysis:
                    type: string
                    example: "Yo, let me break down your financial stack: Current Income: $3000, Current Savings: $500, Monthly Expenses: $2200, Credit Score: 650. Here's what I see: First priority: build that emergency fund to $500. That's your foundation."
                  recommendations:
                    type: object
                    properties:
                      immediate_actions:
                        type: array
                        items:
                          type: string
                        example: ["Build emergency fund", "Pay bills on time", "Track spending"]
                      short_term_goals:
                        type: array
                        items:
                          type: string
                        example: ["Save $500", "Improve credit score", "Start investing $25/month"]
                      long_term_strategies:
                        type: array
                        items:
                          type: string
                        example: ["Build wealth through investments", "Create multiple income streams"]
                  stack_score:
                    type: number
                    format: float
                    example: 7.2
                  motivational_message:
                    type: string
                    example: "Time to stack your bread and build your future! 💰"
              examples:
                analysis_result:
                  summary: Financial analysis result
                  value:
                    user_id: "user123"
                    analysis: "Yo, let me break down your financial stack: Current Income: $3000, Current Savings: $500, Monthly Expenses: $2200, Credit Score: 650. Here's what I see: First priority: build that emergency fund to $500. That's your foundation."
                    recommendations:
                      immediate_actions: ["Build emergency fund", "Pay bills on time", "Track spending"]
                      short_term_goals: ["Save $500", "Improve credit score", "Start investing $25/month"]
                      long_term_strategies: ["Build wealth through investments", "Create multiple income streams"]
                    stack_score: 7.2
                    motivational_message: "Time to stack your bread and build your future! 💰"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /investment/advice:
    post:
      tags:
        - Investment
      summary: Get investment advice
      description: |
        Get personalized investment advice using the free Qwen AI model.
        Provides culturally-aware investment guidance with actionable steps.
      operationId: getInvestmentAdvice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StackMasterMessage'
            examples:
              investment_question:
                summary: Investment advice request
                value:
                  user_id: "user123"
                  message: "I want to start investing in stocks but I'm scared of losing money"
                  context:
                    risk_tolerance: "conservative"
                    investment_amount: 100
      responses:
        '200':
          description: Investment advice provided successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  advice:
                    type: string
                    example: "Real talk - let's break this down. First step is tracking every dollar. You can't stack what you can't see. Start with the 50/30/20 rule: 50% needs, 30% wants, 20% wealth building."
                  tier:
                    type: string
                    example: "free"
                  ai_model:
                    type: string
                    example: "Qwen/Qwen2.5-7B-Instruct"
                  upgrade_message:
                    type: string
                    example: "Want FinRobot agents? Upgrade to $9.99 beta!"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /credit/building-plan:
    post:
      tags:
        - Credit
      summary: Create credit building plan
      description: |
        Create a personalized credit building plan using the free Qwen AI model.
        Provides step-by-step guidance to improve credit scores.
      operationId: createCreditPlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StackMasterMessage'
            examples:
              credit_question:
                summary: Credit building request
                value:
                  user_id: "user456"
                  message: "My credit score is 580 and I want to get it to 700+"
                  context:
                    current_credit_score: 580
                    monthly_income: 2500
      responses:
        '200':
          description: Credit building plan created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  plan:
                    type: string
                    example: "Your credit score is your power score. Every payment on time is money in the bank. Let's get you to 750+ and unlock real wealth opportunities."
                  tier:
                    type: string
                    example: "free"
                  ai_model:
                    type: string
                    example: "Qwen/Qwen2.5-7B-Instruct"
                  upgrade_message:
                    type: string
                    example: "Want advanced FinRobot credit analysis? Upgrade to $9.99 beta!"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    StackMasterMessage:
      type: object
      required:
        - user_id
        - message
      properties:
        user_id:
          type: string
          description: Unique identifier for the user
          example: "user123"
        message:
          type: string
          description: The message to send to The Stack Master
          example: "I want to start investing but I only have $50 a month. What should I do?"
        context:
          type: object
          description: Additional context about the user's financial situation
          additionalProperties: true
          example:
            income: 3000
            savings: 200
            risk_tolerance: "conservative"
      example:
        user_id: "user123"
        message: "I want to start investing but I only have $50 a month. What should I do?"
        context:
          income: 3000
          savings: 200
          risk_tolerance: "conservative"
    StackMasterResponse:
      type: object
      required:
        - response
        - advice_type
      properties:
        response:
          type: string
          description: The Stack Master's response to the user's message
          example: "Real talk - let's break this down. First step is tracking every dollar. You can't stack what you can't see. Start with the 50/30/20 rule: 50% needs, 30% wants, 20% wealth building."
        advice_type:
          type: string
          enum: [general, investment, credit, savings, community, education]
          description: The type of financial advice provided
          example: "investment"
        actionable_steps:
          type: array
          items:
            type: string
          description: Specific steps the user can take
          example: ["Start with $25/month in an index fund", "Research low-cost ETFs like VTI or SPY", "Set up automatic transfers to your investment account"]
        motivational_message:
          type: string
          description: Motivational message to encourage the user
          example: "Your future self will thank you for starting today!"
      example:
        response: "Real talk - let's break this down. First step is tracking every dollar. You can't stack what you can't see. Start with the 50/30/20 rule: 50% needs, 30% wants, 20% wealth building."
        advice_type: "investment"
        actionable_steps: ["Start with $25/month in an index fund", "Research low-cost ETFs like VTI or SPY", "Set up automatic transfers to your investment account"]
        motivational_message: "Your future self will thank you for starting today!"
...
    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message describing what went wrong
          example: "Error chatting with Stack Master: Connection timeout"
      example:
        detail: "Error chatting with Stack Master: Connection timeout"
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for premium features (not required for free tier)
security:
  - {}  # No authentication required for free tier
  - ApiKeyAuth: []  # API key required for premium features
...